experiments/temporal_stability_Euclidean/S7_ARMADA/requirements.txt
# S7 ARMADA Requirements
# ======================
# IMPORTANT: Requires Python 3.9+ (3.12 recommended)
#
# Check available versions:
#   py --list
#
# Install with:
#   py -3.12 -m pip install -r requirements.txt
#
# Run experiments with:
#   py -3.12 run009_drain_capture.py
#
# NOTE: The default 'python' command may point to Python 3.7 which is too old.
# Always use 'py -3.12' (or py -3.13, py -3.14) to specify a newer version.

# Core API clients (all require Python 3.9+)
anthropic>=0.38.0
openai>=1.0.0
google-generativeai>=0.3.0
# Note: xAI/Grok uses OpenAI-compatible API (openai package with base_url="https://api.x.ai/v1")

# Config & utilities
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Visualization (required for visualize_armada.py)
matplotlib>=3.5.0
numpy>=1.21.0
scipy>=1.9.0

# Visualization (optional - for legacy scripts)
networkx>=2.6.0
seaborn>=0.11.0

# Interactive HTML exports (optional)
plotly>=5.0.0
